+++ src/main/java/com/miromaric/dentalassistant/dao/ToothSideLabelDao.java
@@ -12,8 +12,8 @@ public interface ToothSideLabelDao {
     
     void save(ToothSideLabel toothSideLabel);
     List<ToothSideLabel> getAll();
-    ToothSideLabel getOne(int id);
-    ToothSideLabel update(int id,ToothSideLabel toothSideLabel);
-    ToothSideLabel remove(int id);
+    ToothSideLabel getOne(Long id);
+    ToothSideLabel update(Long id,ToothSideLabel toothSideLabel);
+    ToothSideLabel remove(Long id);
     
 }
