+++ src/main/java/com/miromaric/dentalassistant/service/AppointmentService.java
@@ -1,7 +1,6 @@
 package com.miromaric.dentalassistant.service;
 
 import com.miromaric.dentalassistant.model.Appointment;
-import com.miromaric.dentalassistant.model.Patient;
 import java.util.List;
 
 /**
@@ -11,7 +10,7 @@ import java.util.List;
 public interface AppointmentService {
     void save(Appointment appointment);
     List<Appointment> getAll();
-    Appointment getOne(int id);
-    Appointment update(int id,Appointment appointment);
-    Appointment remove(int id);
+    Appointment getOne(Long id);
+    Appointment update(Long id,Appointment appointment);
+    Appointment remove(Long id);
 }
